refactor: Rename S3Service to StorageService and update references across the codebase
