+++ src/main/java/com/mycodingtest/review/MemoService.java
@@ -1,22 +1,22 @@
 package com.mycodingtest.review;
 
-import com.mycodingtest.storage.S3Service;
+import com.mycodingtest.storage.StorageService;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemoService {
 
-    private final S3Service s3Service;
+    private final StorageService storageService;
 
-    public MemoService(S3Service s3Service) {
-        this.s3Service = s3Service;
+    public MemoService(StorageService storageService) {
+        this.storageService = storageService;
     }
 
     public String getMemoUpdateUrl(String reviewId, Long userId) {
-        return s3Service.getMemoUpdateUrl(reviewId, userId);
+        return storageService.getMemoUpdateUrl(reviewId, userId);
     }
 
     public String getMemoReadUrl(String reviewId, Long userId) {
-        return s3Service.getMemoReadUrl(reviewId, userId);
+        return storageService.getMemoReadUrl(reviewId, userId);
     }
 }
