+++ src/main/java/com/mycodingtest/review/SourceCodeService.java
@@ -1,22 +1,22 @@
 package com.mycodingtest.review;
 
-import com.mycodingtest.storage.S3Service;
+import com.mycodingtest.storage.StorageService;
 import org.springframework.stereotype.Service;
 
 @Service
 public class SourceCodeService {
 
-    private final S3Service s3Service;
+    private final StorageService storageService;
 
-    public SourceCodeService(S3Service s3Service) {
-        this.s3Service = s3Service;
+    public SourceCodeService(StorageService storageService) {
+        this.storageService = storageService;
     }
 
     public String getCodeUpdateUrl(String submissionId, Long userId) {
-        return s3Service.getCodeUpdateUrl(submissionId, userId);
+        return storageService.getCodeUpdateUrl(submissionId, userId);
     }
 
     public String getCodeReadUrl(String submissionId, Long userId) {
-        return s3Service.getCodeReadUrl(submissionId, userId);
+        return storageService.getCodeReadUrl(submissionId, userId);
     }
 }
