+++ src/main/java/com/mycodingtest/storage/dto/UrlResponse.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.storage;
+package com.mycodingtest.storage.dto;
 
 public record UrlResponse(
         String url
