+++ src/main/java/com/mycodingtest/storage/impl/S3Service.java
@@ -1,5 +1,6 @@
-package com.mycodingtest.storage;
+package com.mycodingtest.storage.impl;
 
+import com.mycodingtest.storage.StorageService;
 import io.awspring.cloud.s3.S3Template;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -8,7 +9,7 @@ import java.time.Duration;
 import java.util.List;
 
 @Service
-public class S3Service {
+public class S3Service implements StorageService {
 
     private final S3Template s3Template;
     @Value("${aws.s3.bucketName}")
@@ -20,16 +21,19 @@ public class S3Service {
         this.s3Template = s3Template;
     }
 
+
     public String getCodeReadUrl(String submissionId, Long userId) {
         String key = userId + "/codes/" + submissionId + ".txt";
         return s3Template.createSignedGetURL(bucketName, key, Duration.ofSeconds(urlExpirationSeconds)).toString();
     }
 
+    @Override
     public String getCodeUpdateUrl(String submissionId, Long userId) {
         String key = userId + "/codes/" + submissionId + ".txt";
         return s3Template.createSignedPutURL(bucketName, key, Duration.ofSeconds(urlExpirationSeconds)).toString();
     }
 
+    @Override
     public String getMemoReadUrl(String reviewId, Long userId) {
         String key = userId + "/memos/" + reviewId + ".txt";
         if (s3Template.objectExists(bucketName, key)) {
@@ -38,12 +42,13 @@ public class S3Service {
         return "noMemo";
     }
 
+    @Override
     public String getMemoUpdateUrl(String reviewId, Long userId) {
         String key = userId + "/memos/" + reviewId + ".txt";
         return s3Template.createSignedPutURL(bucketName, key, Duration.ofSeconds(urlExpirationSeconds)).toString();
     }
 
-
+    @Override
     public void deleteCodes(List<String> submissionIdList, Long userId) {
         for (String submissionId : submissionIdList) {
             String key = userId + "/codes/" + submissionId + ".txt";
@@ -51,6 +56,7 @@ public class S3Service {
         }
     }
 
+    @Override
     public void deleteMemo(String reviewId, Long userId) {
         String key = userId + "/memos/" + reviewId + ".txt";
         s3Template.deleteObject(bucketName, key);
